+++ domfilters/make4ht-joincharacters.lua
@@ -56,6 +56,9 @@ local function join_characters(obj,par)
   local is_span = function(next_el)
     return charclasses[get_name(next_el)]
   end
+  local count_children = function(curr)
+    return #curr:get_children()
+  end
   local join_elements = function(el, next_el)
     -- it the following element match, copy it's children to the current element
     for _, child in ipairs(next_el:get_children()) do
@@ -79,8 +82,8 @@ local function join_characters(obj,par)
 
   obj:traverse_elements(function(el)
     -- loop over all elements and test if the current element is in a list of
-    -- processed elements (charclasses)
-    if is_span(el) then
+    -- processed elements (charclasses) and if it doesn't contain children
+    if is_span(el) and count_children(el) < 2 then
       local next_el = get_next(el)
       -- loop over the following elements and test whether they are of the same type
       -- as the current one